refactor(SecurityContextPersistenceFilter): 코드 리팩터링
